Issue-34: Docker Compose Container Support
The docker-compose file is now set to use version 2, allowing us to
build the ctia-ui container directly from within the composition. My
tests of the instructions to build the ctia-ui container using the
provided Dockerfile have failed, however, due to missing dependencies.

This Dockerfile uses the officially supported alpine-node container
as its base, and extends it using code from the officially supported
container for alpine-clojure, so that we have a clean and robust base
container upon which to run the CTIA-UI.

Includes support for a dev container as well as a demo container. The
dev container covers all of the dependencies needed to run the ctia-ui
against http://localhost:3000, and the demo container also creates the
ctia-ui container running on port 8080

Closes #34